payment and store view model grdons removed
